Saving a gig

+++ UpComingGibs/ViewModels/GigFormViewModel.cs
@@ -14,7 +14,7 @@ namespace UpComingGibs.ViewModels
 
         public string Time { get; set; }
 
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
 
         public IEnumerable<Genre> Genres { get; set; }
     }
